Bugfix 1
To reduce zipper noise, this patch fades out when a new filter model (and hence new poly patch) is selected. After a 20 ms fadeout, the patch change was supposed to occur. Instead, delay was used improperly and converted the patcher name message into a bang. This was fixed by using a message to read in the patch name and then delay to change the patch.